Error on fitting the model -- degrees overall coded here
Error in vector.namesmatch(target.stats, names(nw.stats)) :
Length of "target.stats" is 4 but should be 7.

Originally had 6 but should be 7 with all degrees and everything by Y/O
age.  In this way the stats.m vector that has the target stats had 6
values.  The main degrees by age were in a 2x3 matrix (Y/O, 012Cp).  We
figured we would need 5 levels to be specified here and that's why 1 was
missing from the seven.  BUT also realized that to get estimate of
target stats from nodefactor (pers,age), we would need the degrees to
not be specified by age to start.  So, now, revamped and returned
anything related to degrees to an overall vector that was not stratified
by age. In this way, the stats.m vector should have just 4 values,
because the main degrees targetstats would only need 2 values (for 012
Cpartners).

BUT still getting this error where the stats.m vector has four values
(as expected), but the program thinks it should have 7. ???  Wanted to
commit to I have this code for degrees overall before I start making
additional changes to fix this error.