Diagnostics Attempt #2
Updated code to have age and PT degree as independent nodefactor terms
(not the interaction of them). Also tried to do concurrent by age term,
but had trouble with diagnostics of this model (static fine, dynamic
not). Will need to resolve this and next add in differential durations
by age.